rag/requirements.txt
pinecone>=3.0.0
chromadb>=0.4.0
numpy>=1.24.0
tqdm>=4.65.0
python-dotenv>=1.0.0